+++ src/main/java/com/students/demo/Controllers/ControllerForJSON.java
@@ -1,23 +1,22 @@
 package com.students.demo.Controllers;
 
 
-import com.students.demo.dto.MyData;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
+import com.students.demo.MyService;
+import com.students.demo.dto.request.DataRequest;
+import com.students.demo.dto.response.MyData;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
 import java.util.Random;
 
 @RestController
 public class ControllerForJSON {
+    @Autowired
+    MyService service;
     @PostMapping("/someData")
-    public MyData getAndChangeJson(@RequestBody MyData someData) {
-        Random random = new Random();
+    public MyData getAndChangeJson(@RequestBody DataRequest someData) {
         MyData newData = new MyData(someData.getPrice(), someData.getInfo());
-        newData.getInfo().setId(random.nextInt(0,100));
+        newData.getInfo().setId(service.GetRandomId());
         return newData;
     }
-
 }
