+++ src/main/java/com/students/demo/dto/request/RequestInfo.java
@@ -1,0 +1,12 @@
+package com.students.demo.dto.request;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.experimental.FieldDefaults;
+
+@Getter
+@Setter
+public class RequestInfo {
+    String date;
+}
