+++ src/main/java/com/students/demo/dto/response/Info.java
@@ -1,11 +1,11 @@
-package com.students.demo.dto;
+package com.students.demo.dto.response;
 
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.experimental.FieldDefaults;
 
-//@FieldDefaults(level= AccessLevel.PRIVATE)
+@FieldDefaults(level= AccessLevel.PRIVATE)
 @Getter
 @Setter
 public class Info {
